added NoItemsView to the project.

+++ ToDoList/Views/ListView.swift
@@ -11,22 +11,28 @@ struct ListView: View {
     
     @EnvironmentObject var listViewModel : ListViewModel
     var body: some View {
-        List{
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack{
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+            }else{
+                List{
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                    
+                    
+                    
+                }
+                .listStyle(PlainListStyle())
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-            
-            
-            
         }
-        .listStyle(PlainListStyle())
         .navigationBarTitle("ToDo Items 📝")
         .navigationBarItems(
             leading: EditButton(),
